refactor: factory namespace
